pyproject.toml
[project]
name = "projetopop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.4",
    "pdfplumber>=0.11.5",
    "pypdf2>=3.0.1",
    "streamlit>=1.41.1",
]
